+++ src/view/InputDemo.vue
@@ -1,148 +1,10 @@
 <template>
 	<div>
-		<!--    <fieldset>-->
-		<!--      <legend>Default</legend>-->
-		<!--      <g-text-field-functional label="Default" persistent hint="This is persistent hint">-->
-		<!--      </g-text-field-functional>-->
-		<!--      <g-text-field-functional label="Prepend outer && non presistent hint"-->
-		<!--                               hint="This is persistent hint">-->
-		<!--        <template v-slot:prepend-outer >-->
-		<!--          <img src="../assets/delivery/avatar.svg" alt="avatar">-->
-		<!--        </template>-->
-		<!--      </g-text-field-functional>-->
-		<!--      <g-text-field-functional label="Prepend inner"-->
-		<!--                               persistent hint="This is persistent hint">-->
-		<!--        <template v-slot:prepend-inner >-->
-		<!--          <img src="../assets/delivery/avatar.svg" alt="avatar">-->
-		<!--        </template>-->
-		<!--      </g-text-field-functional>-->
-		<!--      <g-text-field-functional label="Has prefix" prefix="prefix"-->
-		<!--                               persistent hint="This is persistent hint">-->
-		<!--        <template v-slot:prepend-inner >-->
-		<!--          <img src="../assets/delivery/avatar.svg" alt="avatar">-->
-		<!--        </template>-->
-		<!--      </g-text-field-functional>-->
-		<!--      <g-text-field-functional label="Has suffix" suffix="###"></g-text-field-functional>-->
-		<!--      <g-text-field-functional label="Append inner">-->
-		<!--        <template v-slot:append-inner  >-->
-		<!--          <img src="../assets/delivery/cancel.svg" alt="avatar">-->
-		<!--        </template>-->
-		<!--      </g-text-field-functional>-->
-		<!--      <g-text-field-functional label="Append outer">-->
-		<!--        <template v-slot:append-outer >-->
-		<!--          <img src="../assets/delivery/cancel.svg" alt="avatar">-->
-		<!--        </template>-->
-		<!--      </g-text-field-functional>-->
-		<!--      <g-text-field-functional label="With counter" counter maxlength="25">-->
 
-		<!--      </g-text-field-functional>-->
-		<!--      <g-text-field-functional label="With rule"-->
-		<!--                               hint="5 characters max!"-->
-		<!--                               persistent-->
-		<!--                               counter-->
-		<!--                               maxlength="5"-->
-		<!--                               :rules="rules"-->
-		<!--                               error-message="Exceeded!">-->
-
-		<!--      </g-text-field-functional>-->
-		<!--    </fieldset>-->
-		<!--    <fieldset>-->
-		<!--      <legend>Solo: placeholder instead of label</legend>-->
-		<!--      <g-text-field-functional placeholder="Solo"-->
-		<!--                               solo-->
-		<!--                               persistent-->
-		<!--                               hint="This is persistent hint">-->
-		<!--      </g-text-field-functional>-->
-		<!--      <g-text-field-functional placeholder="Prepend outer && non presistent hint"-->
-		<!--                               solo-->
-		<!--                               hint="This is persistent hint">-->
-		<!--        <template v-slot:prepend-outer >-->
-		<!--          <img src="../assets/delivery/avatar.svg" alt="avatar">-->
-		<!--        </template>-->
-		<!--      </g-text-field-functional>-->
-		<!--      <g-text-field-functional placeholder="Prepend inner"-->
-		<!--                               solo-->
-		<!--                               persistent hint="This is persistent hint">-->
-		<!--        <template v-slot:prepend-inner >-->
-		<!--          <img src="../assets/delivery/avatar.svg" alt="avatar">-->
-		<!--        </template>-->
-		<!--      </g-text-field-functional>-->
-		<!--      <g-text-field-functional placeholder="Has prefix"-->
-		<!--                               solo-->
-		<!--                               prefix="prefix"-->
-		<!--                               persistent-->
-		<!--                               hint="This is persistent hint">-->
-		<!--        <template v-slot:prepend-inner >-->
-		<!--          <img src="../assets/delivery/avatar.svg" alt="avatar">-->
-		<!--        </template>-->
-		<!--      </g-text-field-functional>-->
-		<!--      <g-text-field-functional placeholder="Has suffix"-->
-		<!--                               solo-->
-		<!--                               suffix="###"></g-text-field-functional>-->
-		<!--      <g-text-field-functional placeholder="Append inner">-->
-		<!--        <template v-slot:append-inner  >-->
-		<!--          <img src="../assets/delivery/cancel.svg" alt="avatar">-->
-		<!--        </template>-->
-		<!--      </g-text-field-functional>-->
-		<!--      <g-text-field-functional placeholder="Append outer"-->
-		<!--                               solo>-->
-		<!--        <template v-slot:append-outer >-->
-		<!--          <img src="../assets/delivery/cancel.svg" alt="avatar">-->
-		<!--        </template>-->
-		<!--      </g-text-field-functional>-->
-		<!--      <g-text-field-functional placeholder="With counter"-->
-		<!--                               solo-->
-		<!--                               counter-->
-		<!--                               maxlength="25">-->
-
-		<!--      </g-text-field-functional>-->
-		<!--      <g-text-field-functional placeholder="With rule"-->
-		<!--                               solo-->
-		<!--                               hint="5 characters max!"-->
-		<!--                               persistent-->
-		<!--                               counter-->
-		<!--                               maxlength="5"-->
-		<!--                               :rules="rules"-->
-		<!--                               error-message="Exceeded!">-->
-
-		<!--      </g-text-field-functional>-->
-		<!--      <g-text-field-functional placeholder="Solo outlined"-->
-		<!--                               solo-->
-		<!--                               outlined-->
-		<!--                               persistent-->
-		<!--                               hint="This is persistent hint">-->
-		<!--      </g-text-field-functional>-->
-		<!--      <g-text-field-functional placeholder="Solo flat"-->
-		<!--                               solo-->
-		<!--                               flat-->
-		<!--                               persistent-->
-		<!--                               hint="This is persistent hint">-->
-		<!--      </g-text-field-functional>-->
-		<!--    </fieldset>-->
 		<fieldset>
-			<!--      <legend>Filled</legend>-->
-			<!--      <g-text-field-functional label="Filled"-->
-			<!--                               filled-->
-			<!--                               persistent-->
-			<!--                               hint="This is persistent hint">-->
-			<!--      </g-text-field-functional>-->
-			<!--      <g-text-field-functional label="Prepend outer && non presistent hint"-->
-			<!--                               filled-->
-			<!--                               hint="This is persistent hint">-->
-			<!--        <template v-slot:prepend-outer >-->
-			<!--          <img src="../assets/delivery/avatar.svg" alt="avatar">-->
-			<!--        </template>-->
-			<!--      </g-text-field-functional>-->
-			<!--      <g-text-field-functional label="Prepend inner"-->
-			<!--                               filled-->
-			<!--                               persistent hint="This is persistent hint">-->
-			<!--        <template v-slot:prepend-inner >-->
-			<!--          <img src="../assets/delivery/avatar.svg" alt="avatar">-->
-			<!--        </template>-->
-			<!--      </g-text-field-functional>-->
-			<g-text-field-functional prefix="Prefix" label="label"></g-text-field-functional>
-			<g-text-field-functional label="Mask" v-model="text2" v-mask="mask" ></g-text-field-functional>
-			<g-text-field-functional label="Has prefix"
+			<g-text-field prefix="Prefix" label="label"></g-text-field>
+			<g-text-field label="Mask" v-model="text2" v-mask="mask" ></g-text-field>
+			<g-text-field label="Has prefix"
 															 filled
 															 clearable
 															 counter="25"
@@ -165,40 +27,7 @@
 				<template v-slot:append-outer>
 					<img src="../assets/delivery/avatar.svg" alt="avatar">
 				</template>
-			</g-text-field-functional>
-			<!--            <g-text-field-functional label="Has suffix"-->
-			<!--                                     filled-->
-			<!--                                     suffix="###"></g-text-field-functional>-->
-			<!--            <g-text-field-functional label="Append inner">-->
-			<!--              <template v-slot:append-inner  >-->
-			<!--                <img src="../assets/delivery/cancel.svg" alt="avatar">-->
-			<!--              </template>-->
-			<!--            </g-text-field-functional>-->
-			<!--            <g-text-field-functional label="Append outer"-->
-			<!--                                     filled>-->
-			<!--              <template v-slot:append-outer >-->
-			<!--                <img src="../assets/delivery/cancel.svg" alt="avatar">-->
-			<!--              </template>-->
-			<!--            </g-text-field-functional>-->
-			<!--            <g-text-field-functional label="With counter"-->
-			<!--                                     filled-->
-			<!--                                     counter-->
-			<!--                                     maxlength="25">-->
-
-			<!--            </g-text-field-functional>-->
-			<!--            <g-text-field-functional label="With rule"-->
-			<!--                                     clearable-->
-			<!--                                     outlined-->
-			<!--                                     rounded-->
-			<!--                                     hint="5 characters max!"-->
-			<!--                                     persistent-->
-			<!--                                     counter-->
-			<!--                                     maxlength="5"-->
-			<!--                                     :rules="[ rules.counter]"-->
-			<!--                                     v-model="text1"-->
-			<!--                                     error-message>-->
-
-			<!--            </g-text-field-functional>-->
+			</g-text-field>
 			<g-text-field-bs
 					label="test Label"
 					placeholder="demo placeholder"
@@ -234,12 +63,16 @@
 </template>
 
 <script>
-  import GTextFieldFunctional from '../components/GInput/GTextFieldFunctional';
   import GTextFieldBs from '../components/GInput/GTextFieldBs';
 	import {mask} from 'vue-the-mask'
+
+  function GTextField() {
+
+  }
+
   export default {
     name: 'InputDemo',
-    components: { GTextFieldBs, GTextFieldFunctional },
+    components: { GTextField, GTextFieldBs },
 		directives: {mask},
     data() {
       return {
